+++ src/pico/codegen/polymorphic.c
@@ -612,7 +612,7 @@ void generate_polymorphic_i(Syntax syn, AddressEnv* env, Target target, Internal
         AsmResult out = build_unary_op(ass, JE, imm32(0), a, point);
         size_t start_pos = get_pos(ass);
 
-        uint32_t* jmp_loc = (uint32_t*)get_instructions(ass).data + out.backlink;
+        int32_t* jmp_loc = (int32_t*)(get_instructions(ass).data + out.backlink);
 
         // ---------- TRUE BRANCH ----------
         // now, generate the code to run (if true)
@@ -628,8 +628,8 @@ void generate_polymorphic_i(Syntax syn, AddressEnv* env, Target target, Internal
         } 
 
         // backlink
-        *jmp_loc = (end_pos - start_pos);
-        jmp_loc = (uint32_t*)(get_instructions(ass).data + out.backlink);
+        *jmp_loc = end_pos - start_pos;
+        jmp_loc = (int32_t*)(get_instructions(ass).data + out.backlink);
         start_pos = get_pos(ass);
 
 
@@ -642,7 +642,7 @@ void generate_polymorphic_i(Syntax syn, AddressEnv* env, Target target, Internal
         if (end_pos - start_pos > INT32_MAX) {
             throw_error(point, mk_string("Jump in conditional too large", a));
         } 
-        *jmp_loc = (uint8_t)(end_pos - start_pos);
+        *jmp_loc = end_pos - start_pos;
         break;
     }
     case SLabels: {
